shuffle types around
