+++ packages/backend-storage/src/index.ts
@@ -1,9 +1,3 @@
 export * from './factory.js';
 export { StorageResources, AmplifyStorageProps } from './construct.js';
-export {
-  RoleAccessBuilder,
-  StorageAccessDefinition,
-  StorageAccessBuilder,
-  StorageAction,
-} from './access_builder.js';
 export * from './types.js';
